Trigger automation from alarm time from phone
Want to tigger an automation at a specific time, defined in a
sensor. (More specifically alarm time from phone)

Need to create a template sensor for current time and current day,
in order to trigger automation at a time defined by a sensor value.
Needed to use template trigger to achieve this.